refactor: flatten ulauncher.ui
