+++ ulauncher/ui/preferences/preferences_window.py
@@ -4,12 +4,12 @@ from typing import Any
 
 from ulauncher import paths
 from ulauncher.gi import Gdk, Gtk
-from ulauncher.ui.windows.preferences.views import BaseView, styled
-from ulauncher.ui.windows.preferences.views.about import AboutView
-from ulauncher.ui.windows.preferences.views.extensions import ExtensionsView
-from ulauncher.ui.windows.preferences.views.help import HelpView
-from ulauncher.ui.windows.preferences.views.preferences import PreferencesView
-from ulauncher.ui.windows.preferences.views.shortcuts import ShortcutsView
+from ulauncher.ui.preferences.views import BaseView, styled
+from ulauncher.ui.preferences.views.about import AboutView
+from ulauncher.ui.preferences.views.extensions import ExtensionsView
+from ulauncher.ui.preferences.views.help import HelpView
+from ulauncher.ui.preferences.views.preferences import PreferencesView
+from ulauncher.ui.preferences.views.shortcuts import ShortcutsView
 from ulauncher.utils.system_theme import SystemThemeWatcher
 
 VIEW_CONFIG: list[tuple[str, type[BaseView]]] = [
