+++ ulauncher/ui/preferences/views/shortcuts.py
@@ -5,10 +5,10 @@ from time import time
 
 from ulauncher.gi import Gtk, Pango
 from ulauncher.modes.shortcuts.shortcuts import Shortcut, Shortcuts
-from ulauncher.ui.windows.preferences import views
-from ulauncher.ui.windows.preferences.utils.ext_utils import autofmt_pango_code_block
-from ulauncher.ui.windows.preferences.utils.sidebar_layout import SidebarItem, SidebarLayout
-from ulauncher.ui.windows.preferences.views import (
+from ulauncher.ui.preferences import views
+from ulauncher.ui.preferences.utils.ext_utils import autofmt_pango_code_block
+from ulauncher.ui.preferences.utils.sidebar_layout import SidebarItem, SidebarLayout
+from ulauncher.ui.preferences.views import (
     DialogLauncher,
     TextArea,
     get_window_for_widget,
